requirements.txt
Flask==1.1.2
opencv-python==3.4.7.28
tensorflow-gpu
keras
numpy
flask_cors==3.0.9
easydict==1.9.0
fastapi==0.61.1
pydantic
uvicorn
python-multipart==0.0.5
torch==1.6.0
torchvision==0.7.0
aiofiles==0.5.0
gdown
glob2
pyyaml
timm